Time: 0 ms (100%), Space: 15.4 MB (91.08%) - LeetHub

+++ 2277-count-equal-and-divisible-pairs-in-an-array/2277-count-equal-and-divisible-pairs-in-an-array.cpp
@@ -6,7 +6,7 @@ public:
 
         for (int i = 0; i < n - 1; i++) {
             for (int j = i + 1; j < n; j++) {
-                if (nums[i] == nums[j] && ((long long)i * j) % k == 0) {
+                if (nums[i] == nums[j] && (i * j) % k == 0) {
                     count++;
                 }
             }
